refactor(web): split RemoteInferenceSectionParts.tsx (#663)
apps/web/src/components/RemoteInferenceSectionParts.tsx was 645 LOC.
Split along the natural feature seams:

- RemoteInferencePipelineSteps.ts: HardwareInfo / PipelineStep types,
  the PIPELINE_STEPS constant, isRemoteStep / getCurrentModel helpers,
  shared inputClass.
- RemoteInferencePipelineCards.tsx: the step-card UI cluster — the
  per-step help popover content, PipelineStepCards, EstimatesExplainer,
  and the ApiKeyRequiredDialog.
- RemoteInferenceFireworksParts.tsx: RemoteProviderIntro and
  FireworksApiKeyField.
- RemoteInferencePyannoteParts.tsx: PyannoteCloudIntro and
  PyannoteApiKeyField.

The original RemoteInferenceSectionParts.tsx file is now a thin
re-export barrel, so the single consumer (RemoteInferenceSection)
keeps its existing import path. No behavior changes.
