+++ CHANGELOG.md
@@ -21,6 +21,7 @@ fresh empty `Unreleased` is left at the top.
 ## Unreleased
 
 ### Internal
+- `apps/web/src/components/RemoteInferenceSectionParts.tsx` was 645 LOC; split along its natural seams — pipeline-step descriptors and helpers into `RemoteInferencePipelineSteps.ts`, the step-card UI cluster (cards + estimates explainer + API-key warning dialog) into `RemoteInferencePipelineCards.tsx`, the Fireworks provider intro and API-key entry into `RemoteInferenceFireworksParts.tsx`, and the pyannote-cloud intro / API-key entry into `RemoteInferencePyannoteParts.tsx`. The original module is now a thin re-export barrel so existing imports keep working. ([#663](https://github.com/brlauuu/podlog/issues/663))
 - `apps/pipeline/app/services/meta_analysis.py` was 565 LOC; split out the per-feed / per-episode / per-speaker / timeline / coverage builders into `meta_analysis_aggregations.py` so the orchestrator (stale helpers, `compute_snapshot`, `upsert_snapshot`, `recompute_and_store`) lives on its own. All previously-importable helpers are still importable from `meta_analysis` for back-compat with existing call sites and tests. ([#662](https://github.com/brlauuu/podlog/issues/662))
 - `huggingface-hub` is now declared as a direct pipeline dependency. Previously it was only present transitively via `pyannote-audio` and `whisperx`; `app/services/pyannote.py` imports from it directly, so pinning the version here protects pipeline boot from a future upstream change. Resolved version unchanged. ([#657](https://github.com/brlauuu/podlog/issues/657))
 - CLAUDE.md repo-structure tree refreshed to match the current on-disk layout — adds `backups/` and `notebooks/`, drops stale agent-metadata entries, fills in new pipeline `api/`/`services/` modules and new web `api/`/`lib/` files, bumps the Alembic migration count to 20. ([#658](https://github.com/brlauuu/podlog/issues/658))
